+++ fortran/primeSieve.f90
@@ -5,9 +5,9 @@ module sieve
 contains
 function primeSieve(num) result(primes) 
     implicit none
-    integer, intent(in) :: num
+    integer(kind=8), intent(in) :: num
     logical, dimension(num) :: primes
-    integer :: i, j, current
+    integer(kind=8) :: i, j, current
 
     primes = .true.
     primes(1) = .false.
@@ -30,11 +30,11 @@ end module sieve
 program hello
     use sieve
     implicit none
-    integer :: max, i
+    integer(kind=8) :: max, i
     logical, dimension(:), allocatable :: temp
 
     character(100) :: arg1
-    integer :: arg1int
+    integer(kind=8) :: arg1int
 
     max = 42
     if(command_argument_count() > 0) then
